src/containers/PrivateRouter/index.tsx
import React from 'react';
import { Route, Redirect, RouteProps } from 'react-router-dom';
import auth from '../../services/Auth';

const PrivateRoute: React.FC<RouteProps> = ({ children, ...rest }) => {
  return (
    <Route
      {...rest}
      render={({ location }) =>
        auth.isAuthenticated ? (
          children
        ) : (
          <Redirect
            to={{
              pathname: '/login',
              state: { from: location },
            }}
          />
        )
      }
    />
  );
};

export default PrivateRoute;
